refactor(web): split review sqid encoder into a light module, document admin short codes
Address review feedback on #632 (Codex P1 + P2):

- Extract REVIEW_SQID_ALPHABET/reviewSqid into course-review-sqid.ts so
  the lazy-loaded admin routes import only the sqids encoder; the
  share-image module (html-to-image, avatar, export helpers) is no
  longer bundled into CourseManagementPage/CourseReviewModerationPage
  chunks (verified clean in dist output). course-review-share.ts
  re-exports the two symbols so the detail page and share cards keep
  their existing imports.
- Document in docs/product/current-state.md that admin course-review
  surfaces display sqid short codes consistent with the detail page
  and YourTJCourse-Serverless (issue #628).
